+++ src/Unosquare.Swan/Networking/Ldap/LdapModifyrequest.cs
@@ -1,10 +1,8 @@
 ﻿#if !UWP
 namespace Unosquare.Swan.Networking.Ldap
 {
-    using System;
-
     /// <summary>
-    /// Modification Request
+    /// Represents a LDAP Modification Request Message.
     /// </summary>
     /// <seealso cref="Unosquare.Swan.Networking.Ldap.LdapMessage" />
     public sealed class LdapModifyRequest : LdapMessage
@@ -27,58 +25,6 @@ namespace Unosquare.Swan.Networking.Ldap
         /// The dn.
         /// </value>
         public string DN => Asn1Object.RequestDn;
-
-        /// <summary>
-        /// Gets the modifications.
-        /// </summary>
-        /// <value>
-        /// The modifications.
-        /// </value>
-        /// <exception cref="Exception">Modification Exception</exception>
-        public LdapModification[] Modifications
-        {
-            get
-            {
-                var req = (RfcModifyRequest)Asn1Object.GetRequest();
-
-                var seqof = req.Modifications;
-                var mods = seqof.ToArray();
-                var modifications = new LdapModification[mods.Length];
-
-                for (var m = 0; m < mods.Length; m++)
-                {
-                    var sequence = (Asn1Sequence)mods[m];
-
-                    if (sequence.Size() != 2)
-                    {
-                        throw new InvalidOperationException($"LdapModifyRequest: modification {m} is wrong size:{sequence.Size()}");
-                    }
-
-                    // Contains operation and sequence for the attribute
-                    var operators = sequence.ToArray();
-                    
-                    // get the operation
-                    var op = ((Asn1Enumerated)operators[0]).IntValue();
-                    var attrSeq = (Asn1Sequence)operators[1];
-                    var attrArray = attrSeq.ToArray();
-                    var aname = (Asn1OctetString)attrArray[0];
-                    var name = aname.StringValue();
-                    var avalue = (Asn1SetOf)attrArray[1];
-                    var valueArray = avalue.ToArray();
-                    var attr = new LdapAttribute(name);
-
-                    foreach (var o in valueArray)
-                    {
-                        var t = (Asn1OctetString) o;
-                        attr.AddValue(t.ByteValue());
-                    }
-
-                    modifications[m] = new LdapModification((LdapModificationOp) op, attr);
-                }
-
-                return modifications;
-            }
-        }
 
         /// <inheritdoc />
         public override string ToString() => Asn1Object.ToString();
